+++ strawberryfields/configuration.py
@@ -1,4 +1,4 @@
-# Copyright 2019 Xanadu Quantum Technologies Inc.
+# Copyright 2019-2020 Xanadu Quantum Technologies Inc.
 
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,141 +12,213 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 r"""
-This module contains the :class:`Configuration` class, which is used to
-load, store, save, and modify configuration options for Strawberry Fields.
+This module contains functions used to load, store, save, and modify
+configuration options for Strawberry Fields.
+
+.. warning::
+
+    See more details regarding Strawberry Fields configuration and available
+    configuration options on the :doc:`/introduction/configuration` page.
+
 """
-import os
 import logging as log
+import os
 
 import toml
 from appdirs import user_config_dir
 
 log.getLogger()
 
-
-DEFAULT_CONFIG = {
-    "api": {
-        "authentication_token": "",
-        "hostname": "localhost",
-        "use_ssl": True,
-        "port": 443,
-        "debug": False}
+DEFAULT_CONFIG_SPEC = {
+   "api": {
+        "authentication_token": (str, ""),
+        "hostname": (str, "localhost"),
+        "use_ssl": (bool, True),
+        "port": (int, 443),
+    }
 }
-
-BOOLEAN_KEYS = ("debug", "use_ssl")
-
-
-def parse_environment_variable(key, value):
-    trues = (True, "true", "True", "TRUE", "1", 1)
-    falses = (False, "false", "False", "FALSE", "0", 0)
-
-    if key in BOOLEAN_KEYS:
-        if value in trues:
-            return True
-        elif value in falses:
-            return False
-        else:
-            raise ValueError("Boolean could not be parsed")
-    else:
-        return value
-
 
 class ConfigurationError(Exception):
     """Exception used for configuration errors"""
 
 
-class Configuration:
-    """Configuration class.
+def load_config(filename="config.toml", **kwargs):
+    """Load configuration from keyword arguments, configuration file or
+    environment variables.
 
-    This class is responsible for loading, saving, and storing StrawberryFields
-    and plugin/device configurations.
+    .. note::
+
+        The configuration dictionary will be created based on the following
+        (order defines the importance, going from most important to least
+        important):
+
+        1. keyword arguments passed to ``load_config``
+        2. data contained in environmental variables (if any)
+        3. data contained in a configuration file (if exists)
+
+    Kwargs:
+        filename (str): the name of the configuration file to look for
+
+        Additional configuration options are detailed in
+        :doc:`/introduction/configuration`
+
+    Returns:
+        dict[str, dict[str, Union[str, bool, int]]]: the configuration
+    """
+    config = create_config()
+
+    config_filepath = get_config_filepath(filename=filename)
+
+    if config_filepath is not None:
+        loaded_config = load_config_file(config_filepath)
+        valid_api_options = keep_valid_options(loaded_config["api"])
+        config["api"].update(valid_api_options)
+    else:
+        log.info("No Strawberry Fields configuration file found.")
+
+    update_from_environment_variables(config)
+
+    valid_kwargs_config = keep_valid_options(kwargs)
+    config["api"].update(valid_kwargs_config)
+
+    return config
+
+def create_config(authentication_token="", **kwargs):
+    """Create a configuration object that stores configuration related data
+    organized into sections.
+
+    The configuration object contains API-related configuration options. This
+    function takes into consideration only pre-defined options.
+
+    If called without passing any keyword arguments, then a default
+    configuration object is created.
+
+    Kwargs:
+        Configuration options as detailed in :doc:`/introduction/configuration`
+
+    Returns:
+        dict[str, dict[str, Union[str, bool, int]]]: the configuration
+            object
+    """
+    hostname = kwargs.get("hostname", "localhost")
+    use_ssl = kwargs.get("use_ssl", DEFAULT_CONFIG_SPEC["api"]["use_ssl"][1])
+    port = kwargs.get("port", DEFAULT_CONFIG_SPEC["api"]["port"][1])
+
+    config = {
+        "api": {
+            "authentication_token": authentication_token,
+            "hostname": hostname,
+            "use_ssl": use_ssl,
+            "port": port,
+            }
+    }
+    return config
+
+def get_config_filepath(filename="config.toml"):
+    """Get the filepath of the first configuration file found from the defined
+    configuration directories (if any).
+
+    .. note::
+
+        The following directories are checked (in the following order):
+
+        * The current working directory
+        * The directory specified by the environment variable SF_CONF (if specified)
+        * The user configuration directory (if specified)
+
+    Kwargs:
+        filename (str): the configuration file to look for
+
+    Returns:
+         Union[str, None]: the filepath to the configuration file or None, if
+             no file was found
+    """
+    current_dir = os.getcwd()
+    sf_env_config_dir = os.environ.get("SF_CONF", "")
+    sf_user_config_dir = user_config_dir("strawberryfields", "Xanadu")
+
+    directories = [current_dir, sf_env_config_dir, sf_user_config_dir]
+    for directory in directories:
+        filepath = os.path.join(directory, filename)
+        if os.path.exists(filepath):
+            return filepath
+
+def load_config_file(filepath):
+    """Load a configuration object from a TOML formatted file.
 
     Args:
-        name (str): filename of the configuration file.
-        This should be a valid TOML file. You may also pass an absolute
-        or a relative file path to the configuration file.
+        filepath (str): path to the configuration file
+
+    Returns:
+         dict[str, dict[str, Union[str, bool, int]]]: the configuration
+            object that was loaded
     """
+    with open(filepath, "r") as f:
+        config_from_file = toml.load(f)
+    return config_from_file
 
-    def __str__(self):
-        return "{}".format(self._config)
+def keep_valid_options(sectionconfig):
+    """Filters the valid options in a section of a configuration dictionary.
 
-    def __repr__(self):
-        return "Strawberry Fields Configuration <{}>".format(self._filepath)
+    Args:
+        sectionconfig (dict[str, Union[str, bool, int]]): the section of the
+            configuration to check
 
-    def __init__(self, name="config.toml"):
-        # Look for an existing configuration file
-        self._config = DEFAULT_CONFIG
-        self._config_file = {}
-        self._filepath = None
-        self._name = name
-        self._user_config_dir = user_config_dir("strawberryfields", "Xanadu")
-        self._env_config_dir = os.environ.get("SF_CONF", "")
+    Returns:
+        dict[str, Union[str, bool, int]]: the keep section of the
+            configuration
+    """
+    return {k: v for k, v in sectionconfig.items() if k in VALID_KEYS}
 
-        # Search the current directory, the directory under environment
-        # variable SF_CONF, and default user config directory, in that order.
-        directories = [os.getcwd(), self._env_config_dir, self._user_config_dir]
-        for directory in directories:
-            self._filepath = os.path.join(directory, self._name)
-            try:
-                config = self.load(self._filepath)
-                break
-            except FileNotFoundError:
-                config = False
+def update_from_environment_variables(config):
+    """Updates the current configuration object from data stored in environment
+    variables.
 
-        if config:
-            self.update_config()
-        else:
-            log.info("No Strawberry Fields configuration file found.")
+    The list of environment variables can be found at :mod:`strawberryfields.configuration`
 
-    def update_config(self):
-        """Updates the configuration from either a loaded configuration
-        file, or from an environment variable.
+    Args:
+        config (dict[str, dict[str, Union[str, bool, int]]]): the
+            configuration to be updated
+    Returns:
+        dict[str, dict[str, Union[str, bool, int]]]): the updated
+        configuration
+    """
+    for section, sectionconfig in config.items():
+        env_prefix = "SF_{}_".format(section.upper())
+        for key in sectionconfig:
+            env = env_prefix + key.upper()
+            if env in os.environ:
+                config[section][key] = parse_environment_variable(key, os.environ[env])
 
-        The environment variable takes precedence."""
-        for section, section_config in self._config.items():
-            env_prefix = "SF_{}_".format(section.upper())
+def parse_environment_variable(key, value):
+    """Parse a value stored in an environment variable.
 
-            for key in section_config:
-                # Environment variables take precedence
-                env = env_prefix + key.upper()
+    Args:
+        key (str): the name of the environment variable
+        value (Union[str, bool, int]): the value obtained from the environment
+            variable
 
-                if env in os.environ:
-                    # Update from environment variable
-                    self._config[section][key] = parse_environment_variable(env, os.environ[env])
-                elif self._config_file and key in self._config_file[section]:
-                    # Update from configuration file
-                    self._config[section][key] = self._config_file[section][key]
+    Returns:
+        [str, bool, int]: the parsed value
+    """
+    trues = (True, "true", "True", "TRUE", "1", 1)
+    falses = (False, "false", "False", "FALSE", "0", 0)
 
-    def __getattr__(self, section):
-        if section in self._config:
-            return self._config[section]
+    if DEFAULT_CONFIG_SPEC["api"][key][0] is bool:
+        if value in trues:
+            return True
 
-        raise ConfigurationError("Unknown Strawberry Fields configuration section.")
+        if value in falses:
+            return False
 
-    @property
-    def path(self):
-        """Return the path of the loaded configuration file.
+        raise ValueError("Boolean could not be parsed")
 
-        Returns:
-            str: If no configuration is loaded, this returns ``None``."""
-        return self._filepath
+    if DEFAULT_CONFIG_SPEC["api"][key][0] is int:
+        return int(value)
 
-    def load(self, filepath):
-        """Load a configuration file.
+    return value
 
-        Args:
-            filepath (str): path to the configuration file
-        """
-        with open(filepath, "r") as f:
-            self._config_file = toml.load(f)
-
-        return self._config_file
-
-    def save(self, filepath):
-        """Save a configuration file.
-
-        Args:
-            filepath (str): path to the configuration file
-        """
-        with open(filepath, "w") as f:
-            toml.dump(self._config, f)
+VALID_KEYS = set(create_config()["api"].keys())
+DEFAULT_CONFIG = create_config()
+configuration = load_config()
+config_filepath = get_config_filepath()