+++ .github/CHANGELOG.md
@@ -28,6 +28,11 @@
   `Circuits._apply_gate`, since they were no longer used.
   [#293](https://github.com/XanaduAI/strawberryfields/pull/293/)
 
+* Replaced the `Configuration` class with the `load_config` and auxiliary
+  functions to load configuration from keyword arguments, environment variables
+  and configuration file.
+  [#298](https://github.com/XanaduAI/strawberryfields/pull/298)
+
 ### Bug fixes
 
 * Symbolic Operation parameters are now compatible with TensorFlow 2.0 objects.
@@ -45,7 +50,7 @@
 
 This release contains contributions from (in alphabetical order):
 
-Ville Bergholm, Jack Ceroni, Theodor Isacsson
+Ville Bergholm, Jack Ceroni, Theodor Isacsson, Antal Száva
 
 ---